chore(algebra/{group_power,order}): split off field lemmas
I want to refer to the rational numbers in the definition of a field,
meaning we can't have `algebra.field.basic` in the transitive imports
of `data.rat.basic`.

This is the first step in rearranging those imports: remove the definition of a
field from the dependencies of basic lemmas about `nsmul` and `zsmul`.