feat(geometry/euclidean/basic): `continuous_at_angle`
Add lemmas that (unoriented) angles are continuous, as a function of a
pair of vectors or a triple of points, except where one of the vectors
is zero or one of the end points equals the middle point.

---

Oriented angles will have a corresponding lemma added in a separate PR
once #14980 is in.