feat(analysis/calculus/mean_value): remove assumption in strict_mono_on.strict_convex_on_of_deriv (#15133)
Currently, the lemma `strict_mono_on.strict_convex_on_of_deriv` states that, if a real function `f` is continuous on a convex set `D`, differentiable on its interior, and `deriv f` is strictly monotone on its interior, then `f` is convex on `D`. We remove the differentiability assumption: since `deriv f` is strictly monotone, there is at most one point of nondifferentiability (as `deriv f x = 0` when `f` is not differentiable), and the result is still true (although the proof is a little bit more complicated) in this case.

Of course, in essentially all applications the functions will be differentiable, but the lemma becomes easier to use as the user doesn't need to prove this differentiability.