refactor(tactic/lint/type_classes): change inhabited linter to nonempty linter
The `inhabited` typeclass is intended to be a default value (and computable, if one is programming) but many times it is filled in with a junk (and perhaps noncomputable) value to satisfy the inhabited linter. This switches to pushing contributors to supply a `nonempty` instance. The linter still accepts `inhabited` and `unique` instances.