refactor(data/dfinsupp/basic): Improve definitional equalities of coercions (#15521)
This means that `dfinsupp.coe_add` etc are true by definition, rather than requiring the application of `quotient.induction` first.
The key change is that the underlying function is no longer "hidden" under the quotient, as it does not need to be.

One motivation for this is to make the API more similar to that of `finsupp`.

This change eliminates `dfinsupp.pre`, instead using `{s : multiset ι // ∀ i, i ∈ s ∨ to_fun i = 0}` directly.
We no longer even need to create a `setoid` instance, since we can just use `trunc`.

While adjusting some proofs in `data/finsupp/interval` to use the new definition, this ended up eliminating some decidable arguments. I don't think that is a consequence of this redefinition, and is incidental cleanup that could have been performed separately.