feat(data/sign): `sign_neg_eq_neg_sign`
Add lemmas that the sign of `-a` is the negation of the sign of `a`,
for two kinds of ordered additive groups (corresponding to the type
classes on the `left` and `right` variants of the `neg_pos_iff` and
`neg_neg_iff` lemmas used).

---

I didn't do anything about inconsistent naming of lemmas between the
sign function in this file and other sign functions.  For both
`int.sign` and `real.sign` the equivalent lemma is `sign_neg`, but
here `sign_neg` gives the sign of a negative value, equivalent to
`real.sign_of_neg`.