feat(algebra/order/archimedean): `archimedean_iff_int_lt`, `archimedean_iff_int_le`, `floor_ring.archimedean`
Add lemmas `archimedean_iff_int_lt` and `archimedean_iff_int_le`,
analogous to existing lemmas for `nat` and `rat`.  Deduce that a
linear ordered field that is a floor ring is archimedean (it seems we
currently only have the converse direction).

---

I made `floor_ring.archimedean` a lemma, but I don't know if it should
in fact be an instance, or whether that would cause problems.