refactor(set_theory/ordinal/cantor_normal_form): simplify `CNF` definition (#15449)
We simplify the definition of `CNF` by removing the casing on `b = 0`.

Note that the new definition is not equivalent to the old one: we now have `CNF 0 o = [(0, o)]` instead of `CNF 0 o = []`. This is a good thing though, since it means that `CNF_foldr` (arguably the defining characteristic of the CNF) is now unconditionally true.

We generalize previously existing theorems according to this new definition. Note that none of the theorems have had their hypotheses strengthened, but a few have been weakened.