feat(*): bump to lean 3.46.0 (#15994)
The only relevant change is that `is_strict_total_order` had a redundant assumption removed, and is now exactly the same as `is_strict_total_order'`. This old assumption is proved in `order/rel_classes.lean`, which has to be added as an import to a file, leading to two non-terminal `simp`s breaking.

Of course, the idea is to remove `is_strict_total_order'` in a follow-up PR.